intonation_synthesis/requirements.txt
sagemaker-containers
keras-tcn
numpy>=1.17
setuptools>=41.0.0
nnmnkwii
matplotlib
pandas
scikit-learn
scipy
sklearn
tensorboard>=1.15.0,<1.16.0
tensorflow-estimator>=1.15.0,<1.16.0
tensorflow-gpu>=1.15.0,<1.16.0
seaborn
kerassurgeon
dtaidistance
flask-cors
opencv-python
requests